requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
google-cloud-vision
python-dotenv
Pillow
opencv-python-headless
